+++ models/BaseRedisBlameableModel.php
@@ -5,9 +5,9 @@
  * | | / // // ___//_  _//   ||  __||_   _|
  * | |/ // /(__  )  / / / /| || |     | |
  * |___//_//____/  /_/ /_/ |_||_|     |_|
- * @link http://vistart.name/
+ * @link https://vistart.name/
  * @copyright Copyright (c) 2016 vistart
- * @license http://vistart.name/license/
+ * @license https://vistart.name/license/
  */
 
 namespace vistart\Models\models;
@@ -22,4 +22,30 @@ use vistart\Models\traits\BlameableTrait;
 class BaseRedisBlameableModel extends BaseRedisEntityModel
 {
     use BlameableTrait;
+
+    /**
+     * Initialize the blameable model.
+     * If query class is not specified, [[BaseBlameableQuery]] will be taken.
+     */
+    public function init()
+    {
+        if (!is_string($this->queryClass)) {
+            $this->queryClass = \vistart\Models\queries\BaseRedisBlameableQuery::className();
+        }
+        if ($this->skipInit) {
+            return;
+        }
+        $this->initBlameableEvents();
+        parent::init();
+    }
+
+    /**
+     * Get the query class with specified identity.
+     * @param \vistart\Models\models\BaseUserModel $identity
+     * @return \vistart\Models\queries\BaseBlameableQuery
+     */
+    public static function findByIdentity($identity = null)
+    {
+        return static::find()->byIdentity($identity);
+    }
 }